+++ src/datasources/alarms/query-type-handlers/list-alarms/ListAlarmsQueryHandler.ts
@@ -27,8 +27,9 @@ export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
   public async runQuery(query: ListAlarmsQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
     let mappedFields: DataFrameDTO['fields'] | undefined;
 
-    query.filter = this.transformAlarmsQuery(options.scopedVars, query.filter);
-    const alarmsResponse = await this.queryAlarmsData(query);
+    if (this.isTakeValid(query.take) && this.isPropertiesValid(query.properties)) {
+      query.filter = this.transformAlarmsQuery(options.scopedVars, query.filter);
+      const alarmsResponse = await this.queryAlarmsData(query);
 
     if (this.isPropertiesValid(query.properties)) {
       const alarmsToProcess =
@@ -85,6 +86,8 @@ export class ListAlarmsQueryHandler extends AlarmsQueryHandlerCore {
   private async queryAlarmsData(alarmsQuery: ListAlarmsQuery): Promise<Alarm[]> {
     const alarmsRequestBody: QueryAlarmsRequest = {
       filter: alarmsQuery.filter ?? '',
+      take: alarmsQuery.take,
+      orderByDescending: alarmsQuery.descending ?? DEFAULT_QUERY_EDITOR_DESCENDING,
       transitionInclusionOption: alarmsQuery.transitionInclusionOption ?? DEFAULT_QUERY_EDITOR_TRANSITION_INCLUSION_OPTION,
     }