+++ src/datasources/data-frame/components/v2/DataFrameQueryEditorV2.test.tsx
@@ -84,8 +84,8 @@ const renderComponent = (
         ),
         queryTables: jest.fn().mockResolvedValue(
             [
-                { id: 'table1', name: 'Table 1' },
-                { id: 'table2', name: 'Table 2' },
+                { id: 'table1', name: 'Table 1', columns: [{ name: 'ColumnA' }, {name: 'ColumnB'}] },
+                { id: 'table2', name: 'Table 2', columns: [{ name: 'ColumnD' }, {name: 'ColumnE'}] },
             ]
         ),
     } as unknown as DataFrameDataSource;
@@ -135,7 +135,7 @@ describe("DataFrameQueryEditorV2", () => {
 
     it("should update the query type when a different option is selected", async () => {
         const user = userEvent.setup();
-        const { onChange, onRunQuery } = renderComponent();
+        const { onChange } = renderComponent();
 
         await user.click(screen.getByRole("radio", { name: DataFrameQueryType.Properties }));
 
@@ -144,12 +144,27 @@ describe("DataFrameQueryEditorV2", () => {
                 type: DataFrameQueryType.Properties,
             }));
         });
-        expect(onRunQuery).not.toHaveBeenCalled();
+    });
+
+    it("should call onRunQuery when the query type is changed", async () => {
+        const user = userEvent.setup();
+        const { onRunQuery } = renderComponent();
+
+        await user.click(screen.getByRole("radio", { name: DataFrameQueryType.Properties }));
+
+        expect(onRunQuery).toHaveBeenCalled();
     });
 
     describe("when the query type is data", () => {
         let onChange: jest.Mock;
         let onRunQuery: jest.Mock;
+        
+        beforeAll(() => { 
+            // JSDOM provides offsetHeight as 0 by default. 
+            // Mocking it to return 30 because the ComboBox virtualization relies on this value 
+            // to correctly calculate and render the dropdown options. 
+            jest.spyOn(HTMLElement.prototype, 'offsetHeight', 'get').mockReturnValue(30); 
+        });
 
         beforeEach(() => {
             const result = renderComponent({ type: DataFrameQueryType.Data });
@@ -200,6 +215,42 @@ describe("DataFrameQueryEditorV2", () => {
                     expect(columnsField).toBeInTheDocument();
                     expect(columnsField).toHaveAttribute('aria-expanded', 'false');
                     expect(columnsField).toHaveDisplayValue('');
+                });
+
+                it('should load columns combobox options when filter changes', async () => {
+                    const filterInput = screen.getByTestId("filter-input");
+                    const user = userEvent.setup();
+
+                    await user.clear(filterInput);
+                    await user.type(filterInput, "new filter");
+
+                    await waitFor(() => {
+                        expect(onChange).toHaveBeenCalledWith(expect.objectContaining({
+                            dataTableFilter: "new filter",
+                        }));
+                    });
+
+                    // Click on column combobox to load options
+                    const columnsCombobox = screen.getAllByRole('combobox')[0];
+                    await userEvent.click(columnsCombobox);
+
+                    // Find all option controls by role 'option'
+                    const optionControls = within(document.body).getAllByRole('option');
+                    const optionTexts = optionControls.map(opt => opt.textContent);
+                    expect(optionTexts).toEqual(expect.arrayContaining(
+                        ['ColumnA', 'ColumnB', 'ColumnD', 'ColumnE']
+                    ));
+                });
+
+                it('should not load column options when filter is empty', async () => {
+                    renderComponent({ dataTableFilter: "" });
+                
+                    // Click on column combobox to load options
+                    const columnsCombobox = screen.getAllByRole('combobox')[0];
+                    await userEvent.click(columnsCombobox);
+                
+                    // Assert that no options are shown
+                    expect(within(document.body).queryAllByRole('option').length).toBe(0);
                 });
             });
 
@@ -343,7 +394,9 @@ describe("DataFrameQueryEditorV2", () => {
         let onRunQuery: jest.Mock;
 
         beforeEach(() => {
-            ({ renderResult, onChange, onRunQuery } = renderComponent({ type: DataFrameQueryType.Properties }));
+            ({ renderResult, onChange, onRunQuery } = renderComponent(
+                { type: DataFrameQueryType.Properties }
+            ));
         });
 
         describe("hidden fields", () => {
@@ -383,13 +436,29 @@ describe("DataFrameQueryEditorV2", () => {
 
             it('should call onChange with data table properties when user selects properties', async () => {
                 await userEvent.click(dataTablePropertiesField);
-                await select(dataTablePropertiesField, DataTableProjectionLabelLookup.Properties.label, { container: document.body });
+                await select(
+                    dataTablePropertiesField,
+                    DataTableProjectionLabelLookup.Properties.label,
+                    { container: document.body }
+                );
 
                 await waitFor(() => {
                     expect(onChange).toHaveBeenCalledWith(expect.objectContaining({
                         dataTableProperties: expect.arrayContaining([DataTableProperties.Properties])
                     }));
-                    expect(onRunQuery).not.toHaveBeenCalled();
+                });
+            });
+
+            it('should call onRunQuery when user selects properties', async () => {
+                await userEvent.click(dataTablePropertiesField);
+                await select(
+                    dataTablePropertiesField,
+                    DataTableProjectionLabelLookup.Properties.label,
+                    { container: document.body }
+                );
+
+                await waitFor(() => {
+                    expect(onRunQuery).toHaveBeenCalled();
                 });
             });
 
@@ -422,13 +491,29 @@ describe("DataFrameQueryEditorV2", () => {
 
             it('should call onChange with columns properties when user selects properties', async () => {
                 await userEvent.click(columnPropertiesField);
-                await select(columnPropertiesField, DataTableProjectionLabelLookup.ColumnType.label, { container: document.body });
+                await select(
+                    columnPropertiesField,
+                    DataTableProjectionLabelLookup.ColumnType.label,
+                    { container: document.body }
+                );
 
                 await waitFor(() => {
                     expect(onChange).toHaveBeenCalledWith(expect.objectContaining({
                         columnProperties: expect.arrayContaining([DataTableProperties.ColumnType])
                     }));
-                    expect(onRunQuery).not.toHaveBeenCalled();
+                });
+            });
+
+            it('should call onRunQuery when user selects properties', async () => {
+                await userEvent.click(columnPropertiesField);
+                await select(
+                    columnPropertiesField,
+                    DataTableProjectionLabelLookup.ColumnType.label,
+                    { container: document.body }
+                );
+
+                await waitFor(() => {
+                    expect(onRunQuery).toHaveBeenCalled();
                 });
             });
 
@@ -479,7 +564,8 @@ describe("DataFrameQueryEditorV2", () => {
                 await user.type(takeInput, "0");
 
                 await waitFor(() => {
-                    expect(screen.getByText("The take value must be greater than or equal to 0.")).toBeInTheDocument();
+                    expect(screen.getByText("The take value must be greater than or equal to 0."))
+                        .toBeInTheDocument();
                 });
             });
 
@@ -488,7 +574,8 @@ describe("DataFrameQueryEditorV2", () => {
                 await user.type(takeInput, "5000");
 
                 await waitFor(() => {
-                    expect(screen.getByText("The take value must be less than or equal to 1000.")).toBeInTheDocument();
+                    expect(screen.getByText("The take value must be less than or equal to 1000."))
+                        .toBeInTheDocument();
                 });
             });
 
@@ -497,8 +584,10 @@ describe("DataFrameQueryEditorV2", () => {
                 await user.type(takeInput, "500");
 
                 await waitFor(() => {
-                    expect(screen.queryByText("The take value must be greater than or equal to 0.")).not.toBeInTheDocument();
-                    expect(screen.queryByText("The take value must be less than or equal to 1000.")).not.toBeInTheDocument();
+                    expect(screen.queryByText("The take value must be greater than or equal to 0."))
+                        .not.toBeInTheDocument();
+                    expect(screen.queryByText("The take value must be less than or equal to 1000."))
+                        .not.toBeInTheDocument();
                 });
             });
 
@@ -510,7 +599,15 @@ describe("DataFrameQueryEditorV2", () => {
                     expect(onChange).toHaveBeenCalledWith(expect.objectContaining({
                         take: 500,
                     }));
-                    expect(onRunQuery).not.toHaveBeenCalled();
+                });
+            });
+
+            it("should call onRunQuery when a valid take value is entered", async () => {
+                await user.clear(takeInput);
+                await user.type(takeInput, "500");
+
+                await waitFor(() => {
+                    expect(onRunQuery).toHaveBeenCalled();
                 });
             });
         });
@@ -571,6 +668,19 @@ describe("DataFrameQueryEditorV2", () => {
                 expect(onChange).toHaveBeenCalledWith(expect.objectContaining({
                     dataTableFilter: "new filter",
                 }));
+            });
+        });
+
+        it("should call onRunQuery when the data table filter is changed in the DataTableQueryBuilder component", async () => {
+            const { onRunQuery } = renderComponent({ dataTableFilter: "" });
+            const filterInput = screen.getByTestId("filter-input");
+            const user = userEvent.setup();
+
+            await user.clear(filterInput);
+            await user.type(filterInput, "new filter");
+
+            await waitFor(() => {
+                expect(onRunQuery).toHaveBeenCalled();
             });
         });
     });