+++ src/datasources/data-frame/components/v2/DataFrameQueryEditorV2.tsx
@@ -8,7 +8,17 @@ import { SelectableValue } from '@grafana/data';
 import { Workspace } from 'core/types';
 import { FloatingError } from 'core/errors';
 import { DataTableQueryBuilderFieldNames } from './constants/DataTableQueryBuilder.constants';
-
+import {
+    errorMessages,
+    INLINE_LABEL_WIDTH,
+    VALUE_FIELD_WIDTH,
+    getValuesInPixels,
+    SECTION_WIDTH,
+    DEFAULT_MARGIN_BOTTOM,
+    labels,
+    placeholders,
+    tooltips,
+} from 'datasources/data-frame/constants/v2/DataFrameQueryEditorV2.constants';
 export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onRunQuery, datasource }: PropsV2) => {
     const migratedQuery = datasource.processQuery(query) as ValidDataFrameQueryV2;
 
@@ -17,6 +27,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
     const [isDecimationSettingsSectionOpen, setIsDecimationSettingsSectionOpen] = useState(true);
     const [recordCountInvalidMessage, setRecordCountInvalidMessage] = useState<string>('');
     const [workspaces, setWorkspaces] = useState<Workspace[]>([]);
+    const [columnOptions, setColumnOptions] = useState<Array<ComboboxOption<string>>>([]);
 
     const getPropertiesOptions = (
         type: DataTableProjectionType
@@ -40,13 +51,41 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
         }, [onChange, onRunQuery]
     );
 
-    const onQueryTypeChange = (queryType: DataFrameQueryType) => {
-        handleQueryChange({ ...migratedQuery, type: queryType }, false);
+    const fetchAndSetColumnOptions = async (filter: string) => {
+        if (!filter) {
+            setColumnOptions([]);
+            return;
+        }
+
+        const tables = await datasource.queryTables(
+            filter,
+            TAKE_LIMIT,
+            [
+                DataTableProjections.Name,
+                DataTableProjections.ColumnName,
+                DataTableProjections.ColumnDataType,
+                DataTableProjections.ColumnType,
+            ]
+        );
+        const columnNames = tables
+            .map(table => table.columns ?? [])
+            .flatMap(columns => {
+                return columns
+                    .filter(column => column.name !== undefined)
+                    .map(column => column.name);
+            });
+        setColumnOptions(columnNames.map(name => ({ label: name, value: name })));
     };
 
-    const onDataTableFilterChange = (event?: Event | React.FormEvent<Element>) => {
+    const onQueryTypeChange = (queryType: DataFrameQueryType) => {
+        handleQueryChange({ ...migratedQuery, type: queryType });
+    };
+
+    const onDataTableFilterChange = async (event?: Event | React.FormEvent<Element>) => {
         if (event) {
-            handleQueryChange({ ...migratedQuery, dataTableFilter: (event as CustomEvent).detail.linq }, false);
+            const dataTableFilter = (event as CustomEvent).detail.linq;
+            handleQueryChange({ ...migratedQuery, dataTableFilter });
+            await fetchAndSetColumnOptions(dataTableFilter);
         }
     };
 
@@ -54,14 +93,14 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
         const dataTableProperties = properties
             .filter(property => property.value !== undefined)
             .map(property => property.value as DataTableProperties);
-        handleQueryChange({ ...migratedQuery, dataTableProperties }, false);
+        handleQueryChange({ ...migratedQuery, dataTableProperties });
     };
 
     const onColumnPropertiesChange = (properties: Array<SelectableValue<DataTableProperties>>) => {
         const columnProperties = properties
             .filter(property => property.value !== undefined)
             .map(property => property.value as DataTableProperties);
-        handleQueryChange({ ...migratedQuery, columnProperties }, false);
+        handleQueryChange({ ...migratedQuery, columnProperties });
     };
 
     const onTakeChange = (event: React.FormEvent<HTMLInputElement>) => {
@@ -69,7 +108,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
         const message = validateTakeValue(value, TAKE_LIMIT);
 
         setRecordCountInvalidMessage(message);
-        handleQueryChange({ ...migratedQuery, take: value }, false);
+        handleQueryChange({ ...migratedQuery, take: value });
     };
 
     const onColumnsChange = (columns: Array<ComboboxOption<string>>) => {
@@ -135,7 +174,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
         <>
             <InlineField
                 label={labels.queryType}
-                labelWidth={inlineLabelWidth}
+                labelWidth={INLINE_LABEL_WIDTH}
                 tooltip={tooltips.queryType}
             >
                 <RadioButtonGroup
@@ -149,12 +188,12 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                 <>
                     <InlineField
                         label={labels.dataTableProperties}
-                        labelWidth={inlineLabelWidth}
+                        labelWidth={INLINE_LABEL_WIDTH}
                         tooltip={tooltips.dataTableProperties}
                     >
                         <MultiSelect
                             placeholder={placeholders.dataTableProperties}
-                            width={valueFieldWidth}
+                            width={VALUE_FIELD_WIDTH}
                             value={migratedQuery.dataTableProperties}
                             onChange={onDataTablePropertiesChange}
                             options={dataTablePropertiesOptions}
@@ -164,12 +203,12 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                     </InlineField>
                     <InlineField
                         label={labels.columnProperties}
-                        labelWidth={inlineLabelWidth}
+                        labelWidth={INLINE_LABEL_WIDTH}
                         tooltip={tooltips.columnProperties}
                     >
                         <MultiSelect
                             placeholder={placeholders.columnProperties}
-                            width={valueFieldWidth}
+                            width={VALUE_FIELD_WIDTH}
                             value={migratedQuery.columnProperties}
                             onChange={onColumnPropertiesChange}
                             options={columnPropertiesOptions}
@@ -181,7 +220,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
             )}
 
             <div
-                style={{ width: getValuesInPixels(sectionWidth) }}
+                style={{ width: getValuesInPixels(SECTION_WIDTH) }}
             >
                 <Collapse
                     label={labels.queryConfigurations}
@@ -190,14 +229,14 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                     onToggle={() => setIsQueryConfigurationSectionOpen(!isQueryConfigurationSectionOpen)}
                 >
                     <InlineLabel
-                        width={valueFieldWidth}
+                        width={VALUE_FIELD_WIDTH}
                         tooltip={tooltips.queryByDataTableProperties}
                     >
                         {labels.queryByDataTableProperties}
                     </InlineLabel>
                     <div style={{
-                        width: getValuesInPixels(valueFieldWidth),
-                        marginBottom: getValuesInPixels(defaultMarginBottom)
+                        width: getValuesInPixels(VALUE_FIELD_WIDTH),
+                        marginBottom: getValuesInPixels(DEFAULT_MARGIN_BOTTOM)
                     }}>
                         <DataTableQueryBuilder
                             filter={migratedQuery.dataTableFilter}
@@ -211,7 +250,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                     {migratedQuery.type === DataFrameQueryType.Properties && (
                         <InlineField
                             label={labels.take}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.take}
                             invalid={!!recordCountInvalidMessage}
                             error={recordCountInvalidMessage}
@@ -233,7 +272,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
 
             {migratedQuery.type === DataFrameQueryType.Data && (
                 <div
-                    style={{ width: getValuesInPixels(sectionWidth) }}
+                    style={{ width: getValuesInPixels(SECTION_WIDTH) }}
                 >
                     <Collapse
                         label={labels.columnConfigurations}
@@ -243,21 +282,22 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                     >
                         <InlineField
                             label={labels.columns}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.columns}
                         >
                             <MultiCombobox
                                 placeholder={placeholders.columns}
-                                width={inlineLabelWidth}
+                                width={40}
                                 value={migratedQuery.columns}
                                 onChange={onColumnsChange}
-                                options={[]}
+                                options={columnOptions}
                                 createCustomValue={false}
+                                isClearable={true}
                             />
                         </InlineField>
                         <InlineField
                             label={labels.includeIndexColumns}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.includeIndexColumns}
                         >
                             <InlineSwitch
@@ -267,7 +307,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                         </InlineField>
                         <InlineField
                             label={labels.filterNulls}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.filterNulls}
                         >
                             <InlineSwitch
@@ -285,11 +325,11 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                     >
                         <InlineField
                             label={labels.decimationMethod}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.decimationMethod}
                         >
                             <Combobox
-                                width={inlineLabelWidth}
+                                width={INLINE_LABEL_WIDTH}
                                 value={migratedQuery.decimationMethod}
                                 onChange={onDecimationMethodChange}
                                 options={decimationMethods}
@@ -298,12 +338,12 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                         </InlineField>
                         <InlineField
                             label={labels.xColumn}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.xColumn}
                         >
                             <Combobox
                                 placeholder={placeholders.xColumn}
-                                width={inlineLabelWidth}
+                                width={INLINE_LABEL_WIDTH}
                                 value={migratedQuery.xColumn}
                                 onChange={onXColumnChange}
                                 options={[]}
@@ -312,7 +352,7 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
                         </InlineField>
                         <InlineField
                             label={labels.useTimeRange}
-                            labelWidth={inlineLabelWidth}
+                            labelWidth={INLINE_LABEL_WIDTH}
                             tooltip={tooltips.useTimeRange}
                         >
                             <InlineSwitch
@@ -331,60 +371,3 @@ export const DataFrameQueryEditorV2: React.FC<PropsV2> = ({ query, onChange, onR
         </>
     );
 };
-
-const labels = {
-    queryType: 'Query type',
-    dataTableProperties: 'Data table properties',
-    columnProperties: 'Column properties',
-    queryConfigurations: 'Query configurations',
-    columnConfigurations: 'Column configurations',
-    decimationSettings: 'Decimation settings',
-    queryByDataTableProperties: 'Query by data table properties',
-    columns: 'Columns',
-    filterNulls: 'Filter nulls',
-    includeIndexColumns: 'Include index columns',
-    decimationMethod: 'Decimation method',
-    xColumn: 'X-column',
-    useTimeRange: 'Use time range',
-    take: 'Take',
-};
-const tooltips = {
-    queryType: 'This field specifies the type for the query that searches the data tables. The query can retrieve row data or metadata.',
-    queryByDataTableProperties: 'This optional field applies a filter to a query while searching the data tables.',
-    take: 'This field sets the maximum number of records to return from the query.',
-    columns: 'Specifies the columns to include in the response data.',
-    filterNulls: `Specifies whether to filter out null and NaN values before decimating the data.`,
-    includeIndexColumns: 'Specifies whether to include index columns in the response data.',
-    dataTableProperties: 'This field specifies the data table properties to be queried.',
-    columnProperties: 'This field specifies the column properties to be queried.',
-    decimationMethod: 'Specifies the method used to decimate the data.',
-    xColumn: 'Specifies the column to use for the x-axis when decimating the data.',
-    useTimeRange: `Specifies whether to query only for data within the dashboard time range if the
-                table index is a timestamp. Enable when interacting with your data on a graph.`,
-};
-const placeholders = {
-    dataTableProperties: 'Select data table properties to fetch',
-    columnProperties: 'Select column properties to fetch',
-    take: 'Enter record count',
-    columns: 'Select columns',
-    xColumn: 'Select x-column',
-};
-
-const errorMessages = {
-    take: {
-        greaterOrEqualToZero: 'The take value must be greater than or equal to 0.',
-        lessOrEqualToTakeLimit: `The take value must be less than or equal to ${TAKE_LIMIT}.`
-    }
-};
-
-const getValuesInPixels = (valueInGrafanaUnits: number) => {
-    return valueInGrafanaUnits * 8 + 'px';
-};
-
-// The following values are multiples of 8 to align with Grafana's grid system, hence 25 in grafana 
-// is equal to 25*8 = 200px.
-const inlineLabelWidth = 25;
-const valueFieldWidth = 65.5;
-const inlineMarginBetweenLabelAndField = 0.5;
-const defaultMarginBottom = 1;
-const sectionWidth = inlineLabelWidth + valueFieldWidth + inlineMarginBetweenLabelAndField;