+++ src/shared/components/ResultsQueryBuilder/ResultsQueryBuilder.tsx
@@ -1,21 +1,16 @@
-import { useTheme2 } from '@grafana/ui';
 import { queryBuilderMessages, QueryBuilderOperations } from 'core/query-builder.constants';
 import { expressionBuilderCallback, expressionReaderCallback } from 'core/query-builder.utils';
 import { Workspace, QueryBuilderOption, QBField } from 'core/types';
 import { filterXSSField, filterXSSLINQExpression } from 'core/utils';
 
 import React, { useState, useEffect, useMemo } from 'react';
-import QueryBuilder, { QueryBuilderCustomOperation, QueryBuilderProps } from 'smart-webcomponents-react/querybuilder';
+import { QueryBuilderCustomOperation, QueryBuilderProps } from 'smart-webcomponents-react/querybuilder';
 
-import 'smart-webcomponents-react/source/styles/smart.dark-orange.css';
-import 'smart-webcomponents-react/source/styles/smart.orange.css';
-import 'smart-webcomponents-react/source/styles/components/smart.base.css';
-import 'smart-webcomponents-react/source/styles/components/smart.common.css';
-import 'smart-webcomponents-react/source/styles/components/smart.querybuilder.css';
 import {
   ResultsQueryBuilderFields,
   ResultsQueryBuilderStaticFields,
 } from 'datasources/results/constants/ResultsQueryBuilder.constants';
+import { SlQueryBuilder } from 'core/components/SlQueryBuilder/SlQueryBuilder';
 
 type ResultsQueryBuilderProps = QueryBuilderProps &
   React.HTMLAttributes<Element> & {
@@ -34,8 +29,6 @@ export const ResultsQueryBuilder: React.FC<ResultsQueryBuilderProps> = ({
   status,
   globalVariableOptions,
 }) => {
-  const theme = useTheme2();
-  document.body.setAttribute('theme', theme.isDark ? 'dark-orange' : 'orange');
 
   const [fields, setFields] = useState<QBField[]>([]);
   const [operations, setOperations] = useState<QueryBuilderCustomOperation[]>([]);
@@ -210,13 +203,12 @@ export const ResultsQueryBuilder: React.FC<ResultsQueryBuilderProps> = ({
   }, [workspaceField, startedAtField, updatedAtField, partNumberField, globalVariableOptions, statusField]);
 
   return (
-    <QueryBuilder
+    <SlQueryBuilder
       customOperations={operations}
       fields={fields}
       messages={queryBuilderMessages}
       onChange={onChange}
       value={sanitizedFilter}
-      fieldsMode="static"
     />
   );
 };